Add first enricher for pair programming activities
This module controls those authors and committers that are
different. If so, those commits are duplicated in the original
data frame with updated info about the original author.

This module works when the process within the organization
uses author and committer as a pair programming process.
For other cases such as identification of pair programming
in the log message or in the author name, further work
is needed.