Add incremental updates support in Areas of Code
Now it is possible to build areas of code index incrementally by
running the builder with `inc=true` in the General section of
`.settings`.

To be able to build the index incrementally, we need to build
a custom `_id`, unique for each event we generate. We use the
following combination to achieve this:
Perceval_UUID + "_" + filepath + "_" + fileaction

Fixed a bug that didn't upload the last bunch of events. The
previous bunch was uploaded twice instead (it just overwrote
those items instead of uploading the new ones).

Some refactoring to get a clear view of the main loop.

Also adds support for merges filter by adding `files` field to
all items. This field will contain the number of files touched
by the given commit. This way we have this number available for
all file events generated for a commit.

Code compliant with PEP8.