Remove grimoire-elk dependency
The dependency has been removed because it isn't
imported directly on cereslib but it uses some
objects related to it. However, if these objects
have the same API, any other library can be used.

This is not ideal but, right now, it's the only
way to remove the cyclic dependency we have between
'cereslib' and 'grimoire-elk' without the use
of a third library that shares the code needed
by both.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>